requirements.txt
fastapi
uvicorn
pandas
Jinja2==3.0.3
fastapi-sqlalchemy
python-dotenv
flask
